ui/server changes

+++ trunk/src/com/kpos/domain/OrderItem.java
@@ -66,7 +66,7 @@ public class OrderItem extends AbstractDomainObject {
         this.saleItem = saleItem;
     }
     
-    public float getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
